Improve vhost switch error handling and lifecycle management; add tests for aborted switches
